dv-5 RSS408-164 - move from SessionFactory to EntityManager
This greatly helps reduce the Spring Config and allows us to use SpringDataJpa interfaces for common database operations.

+++ datavault-common/src/main/java/org/datavaultplatform/common/model/dao/DepositReviewDAO.java
@@ -8,5 +8,5 @@ public interface DepositReviewDAO extends BaseDAO<DepositReview> {
 
     List<DepositReview> search(String query);
 
-    int count();
+    long count();
 }
